% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aic.R
\name{compute_aic}
\alias{compute_aic}
\title{Title}
\usage{
compute_aic(estProp, Y.raw)
}
\arguments{
\item{estProp}{A numeric matrix representing the estimated proportions of each cell type for each sample.
Rows correspond to genes, and columns correspond to samples.}

\item{Y.raw}{A numeric matrix representing the raw gene expression data, where rows correspond to genes and columns to samples.}
}
\value{
A numeric value representing the Akaike Information Criterion (AIC) for the model.
}
\description{
#' Compute Akaike Information Criterion (AIC) for Deconvolution Model
}
\details{
This function computes the Akaike Information Criterion (AIC) for a given deconvolution model
by evaluating the residual sum of squares (RSS) and the number of parameters in the model.
The AIC is used to compare models with different numbers of parameters and select the best-fitting model.
}
\examples{
# Example usage of compute_aic
estProp <- matrix(runif(100), nrow = 50, ncol = 2)  # Example estimated proportions (50 genes x 2 cell types)
Y_raw <- matrix(rnorm(1000), nrow = 50, ncol = 20)  # Example gene expression matrix (50 genes x 20 samples)
aic_value <- compute_aic(estProp, Y_raw)
print(aic_value)
}
