% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csDeconv.R
\name{csDeconv}
\alias{csDeconv}
\title{Title csDeconv}
\usage{
csDeconv(
  Y_raw,
  K,
  nMarker = 1000,
  InitMarker = NULL,
  TotalIter = 30,
  bound_negative = FALSE
)
}
\arguments{
\item{Y_raw}{A numeric matrix or a \code{SummarizedExperiment} object. The rows represent genes
and the columns represent samples. This is the input data from complex tissues for  deconvolution.}

\item{K}{The number of pure cell types to be estimated from the data.}

\item{nMarker}{The number of marker genes to use for the deconvolution. Default is 1000.}

\item{InitMarker}{A vector of marker gene names to be used for the deconvolution. If not
provided, the top variable genes will be selected based on the input data.}

\item{TotalIter}{The total number of iterations to perform for the deconvolution process.
Default is 30.}

\item{bound_negative}{The total number of iterations to perform for the deconvolution process.
Default is 30.}
}
\value{
A list containing the following components:
\item{allRMSE}{A vector of RMSE values for each iteration of the deconvolution process.}
\item{allProp}{A list containing the estimated proportions for each iteration.}
\item{estProp}{The estimated proportions of pure cell types for the sample corresponding
to the iteration with the minimum RMSE.}
\item{updatedInx}{The indices of the marker genes used for the deconvolution.}
}
\description{
#' Cell Type Deconvolution Using Reference-Free Deconvolution Method
}
\details{
This function performs cell type deconvolution on high-throughput tissue measurement data.
The deconvolution process aims to estimate the proportion of pure cell types present in
each sample based on gene expression data from complex tissue samples. The deconvolution
is reference-free, and the function supports iterative refinement to improve proportion
estimates.
}
\examples{
# Example of deconvolution with random data:
Y_raw <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Example data
K <- 3  # Number of cell types
deconv_result <- csDeconv(Y_raw, K, nMarker = 500)

# Access the estimated proportions
deconv_result$estProp
}
