% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEK_wrapper.R
\name{DEK_wrapper}
\alias{DEK_wrapper}
\title{Title DEK_wrapper}
\usage{
DEK_wrapper(K, Y, Prop, design, var.threshold = 0.1)
}
\arguments{
\item{K}{The number of tissues in the dataset.}

\item{Y}{A numeric matrix or a \code{SummarizedExperiment} object representing the expression
matrix. Rows are features (genes), and columns are samples.}

\item{Prop}{A numeric matrix representing the tissue proportions for each sample. Rows are
samples, and columns are tissues.}

\item{design}{A numeric vector specifying the experimental design. If it contains zeros,
pairwise tissue comparisons will be made.}

\item{var.threshold}{A numeric value indicating the threshold for variance filtering (default is 0.1).}
}
\value{
A list of differential expression results. The list contains either:
\itemize{
\item Tissue-specific differential expression results, if \code{design} is not all zeros.
\item Pairwise tissue comparisons, if \code{design} is all zeros.
The results are returned as lists, with names corresponding to either tissue labels or
pairwise tissue comparisons.
}
}
\description{
#' Differential Expression with Tissue Proportions
}
\details{
This function performs differential expression analysis between tissues based on their
proportions in the mixed tissue samples. It uses a wrapper around the \code{DEKTissue} function
to calculate tissue-specific differential expression, considering both single and multiple
tissue comparisons.
}
\examples{
# Example of using DEK_wrapper to perform differential expression:
K <- 3
Y <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Example expression data
Prop <- matrix(runif(30), nrow = 10, ncol = 3)  # Example tissue proportions
design <- c(0, 1, 0)  # Example design vector (e.g., for pairwise comparisons)
DE_results <- DEK_wrapper(K, Y, Prop, design)
}
