% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEKTissue.R
\name{DEKTissue}
\alias{DEKTissue}
\title{Title DEKTissue}
\usage{
DEKTissue(
  K,
  Y,
  Prop,
  design,
  WhichPar = NULL,
  contrast.vec = NULL,
  sort = F,
  var.threshold = 0.1,
  data.threshold = TRUE
)
}
\arguments{
\item{data.threshold}{}
}
\value{
A data frame containing the differential expression analysis results. The results include:
\itemize{
\item \code{t.stat}: The t-statistic for each gene.
\item \code{t.pval}: The p-value for each test.
\item \code{t.fdr}: The adjusted p-value using the false discovery rate (FDR) method.
\item Additional columns for estimated coefficients and effect sizes.
}
}
\description{
#' Differential Expression Analysis for Tissue Proportions
}
\details{
This function performs differential expression analysis considering tissue proportions
in complex tissue samples. It estimates the effect of each tissue type and compares
tissues in terms of their expression profiles, taking into account different experimental
designs and potential contrast vectors.
}
\examples{
# Example of using DEKTissue to perform differential expression analysis:
K <- 3
Y <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Example gene expression data
Prop <- matrix(runif(30), nrow = 10, ncol = 3)  # Example tissue proportions
design <- c(0, 1, 0)  # Example experimental design
res <- DEKTissue(K, Y, Prop, design, WhichPar=1)

# Access the results:
res
}
