% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEVarSelect.R
\name{DEVarSelect}
\alias{DEVarSelect}
\title{Title DEVarSelect}
\usage{
DEVarSelect(Y_raw, Prop0, nMarker = 1000, bound_negative = FALSE)
}
\arguments{
\item{Y_raw}{A numeric matrix or a \code{SummarizedExperiment} object. The rows represent genes
and the columns represent samples. This is the input data from complex tissues for marker
gene selection.}

\item{Prop0}{A numeric matrix of proportions, with rows representing samples and columns
representing cell types. These are the initial estimates of the cell type proportions.}

\item{nMarker}{The number of marker genes to select for each cell type. Default is 1000.}

\item{bound_negative}{A logical value indicating whether to bound the selected proportions
to be non-negative. Default is \code{FALSE}.}
}
\value{
A vector of indices corresponding to the selected marker genes. These genes are
expected to be the most differential across the tissue types and useful for the deconvolution process.
}
\description{
#' Select Differentially Expressed Marker Genes for Deconvolution
}
\details{
This function selects marker genes based on their differential expression across cell types.
It uses a statistical method to identify tissue-specific genes, which are then used in
the deconvolution process. The function selects a set of markers for each cell type
to improve the estimation of cell type proportions in mixed tissue samples.
}
\examples{
# Example of selecting marker genes for deconvolution:
Y_raw <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Example data
Prop0 <- matrix(runif(30), nrow = 10, ncol = 3)  # Example cell type proportions
marker_genes <- DEVarSelect(Y_raw, Prop0, nMarker = 500)

# Access the selected marker genes
marker_genes
}
