% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoCBS.R
\name{DoCBS}
\alias{DoCBS}
\title{Title}
\usage{
DoCBS(Y, W, nu.v = c(0.25, 0.5, 0.75))
}
\arguments{
\item{Y}{A numeric matrix of size N x P, where N is the number of samples and P is the number of target variables (e.g., genes).}

\item{W}{A numeric matrix of size N x Q, where N is the number of samples and Q is the number of basis (e.g., features).}

\item{nu.v}{A numeric vector of values for the \code{nu} parameter in the SVM model, default is \code{c(0.25, 0.5, 0.75)}.
The \code{nu} parameter controls the upper bound on the fraction of margin errors and the lower bound on the fraction of support vectors.}
}
\value{
A numeric matrix of size P x Q, where P is the number of target variables and Q is the number of basis, containing
the estimated coefficients for each target variable and basis after selecting the optimal \code{nu}.
}
\description{
#' Perform CBS (Constrained Basis Selection) Using Support Vector Machines
}
\details{
This function performs CBS (Constrained Basis Selection) using Support Vector Machines (SVM) to select the optimal basis
for reconstructing data. The function estimates the coefficients for each sample using a linear SVM with a specified
\code{nu} parameter, and then selects the best \code{nu} based on the root mean squared error (RMSE) between the observed
and reconstructed values.
}
\examples{
# Example usage of DoCBS
Y <- matrix(rnorm(100), nrow = 10, ncol = 10)  # Example target matrix (10 samples x 10 variables)
W <- matrix(rnorm(100), nrow = 10, ncol = 5)   # Example basis matrix (10 samples x 5 features)
H <- DoCBS(Y, W, nu.v = c(0.25, 0.5, 0.75))
}
