% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellNumber.R
\name{getCellNumber}
\alias{getCellNumber}
\title{Title getCellNumber}
\usage{
getCellNumber(Y.raw, possibleCellNumber = 2:15)
}
\arguments{
\item{Y.raw}{A numeric matrix or a SummarizedExperiment object representing gene expression data.
The rows represent genes (features) and columns represent samples.}

\item{possibleCellNumber}{A numeric vector specifying the possible number of cell types to evaluate, default is 2:15.}
}
\value{
A list with the following elements:
\item{bestK}{The optimal number of cell types based on the lowest AIC value.}
\item{allAIC}{A numeric vector of AIC values for each tested cell number.}
}
\description{
#' Estimate the Optimal Number of Cells Using AIC
}
\details{
This function estimates the optimal number of cell types by applying a deconvolution method to a given gene expression matrix
and comparing the resulting models using the Akaike Information Criterion (AIC). The function iterates over a range of possible
cell numbers and selects the one with the lowest AIC value.
}
\examples{
# Example usage of getCellNumber
Y_raw <- matrix(rnorm(1000), nrow = 50, ncol = 20)  # Example gene expression matrix (50 genes x 20 samples)
result <- getCellNumber(Y.raw = Y_raw, possibleCellNumber = 2:10)
print(result$bestK)
print(result$allAIC)
}
