% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPropAligned.R
\name{GetPropAligned}
\alias{GetPropAligned}
\title{Title GetPropAligned}
\usage{
GetPropAligned(input, reference)
}
\arguments{
\item{input}{A numeric matrix where the columns represent different samples or
features that need to be aligned with the reference matrix.}

\item{reference}{A numeric matrix that serves as the reference for alignment.
The columns of the reference matrix should correspond to those in the input matrix.}
}
\value{
A matrix with the columns of the input matrix reordered to align with
the reference matrix based on correlation.
}
\description{
#' Align Proportions Based on Correlation Coefficient
}
\details{
This function aligns the columns of a given input matrix to a reference matrix based
on the correlation coefficients between their columns. The alignment is performed by
finding the maximum correlation for each column in the input matrix with columns in
the reference matrix. The function returns a transformed input matrix with columns
reordered according to their best correlation matches with the reference.

This method is useful when you have a reference dataset, and you want to align or
reorder another dataset such that the columns of the input matrix match as closely
as possible to the reference matrix based on correlation.
}
\examples{
# Example of aligning an input matrix with a reference matrix based on correlation
input_data <- matrix(rnorm(100), nrow = 10, ncol = 5)  # Random data for illustration
reference_data <- matrix(rnorm(100), nrow = 10, ncol = 5)
aligned_data <- GetPropAligned(input_data, reference_data)
}
