% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{getProportion}
\alias{getProportion}
\title{Title getProportion}
\usage{
getProportion(N_sample, L = 4)
}
\arguments{
\item{N_sample}{An integer specifying the number of samples to generate. Each sample will have a proportion vector.}

\item{L}{An integer specifying the number of categories or components. The default is 4.}
}
\value{
A numeric matrix of size \verb{N_sample x L}, where each row represents a sample and the columns represent the proportions
for each category/component. The values in each row sum to 1, as they represent proportions.
}
\description{
Generate Proportions for Dirichlet Distribution
}
\details{
This function generates proportions for a given number of samples based on a Dirichlet distribution.
The Dirichlet distribution is parameterized using pre-defined concentration parameters (\code{alpha.ctr}) that are
either fixed for 4 components or can be adjusted for a different number of components (\code{L}).
The generated proportions are typically used to model the relative abundance of different categories or tissues.
}
\examples{
#' # Generate proportions for 100 samples with 4 categories
proportions <- getProportion(100)
head(proportions)

# Generate proportions for 50 samples with 6 categories
proportions_6 <- getProportion(50, L = 6)
head(proportions_6)
}
