% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleMix.R
\name{getSampleMix}
\alias{getSampleMix}
\title{Title getSampleMix}
\usage{
getSampleMix(N_sample, logpure_base, logpure_sd, noise_sd = 0.1)
}
\arguments{
\item{N_sample}{An integer specifying the number of mixed samples to generate.}

\item{logpure_base}{A numeric matrix of size \verb{p x K} representing the base values for each pure tissue.
The rows correspond to features (genes), and the columns correspond to tissues.}

\item{logpure_sd}{A numeric matrix of size \verb{p x K} representing the standard deviation for each pure tissue.}

\item{noise_sd}{A numeric value specifying the standard deviation of noise to be added to the generated samples.
The default value is 0.1.}
}
\value{
A list containing:
\item{obs.Y}{A numeric matrix of size \verb{p x N_sample} representing the generated mixed samples, where each row is a feature and each column is a sample.}
\item{trueProp}{A numeric matrix of size \verb{N_sample x K} representing the true proportions of each tissue in each sample.}
\item{tmp}{A matrix containing the individual tissue contributions for each sample before noise is added.}
}
\description{
#' Generate Mixed Samples Based on Pure Tissue References
}
\details{
This function generates mixed samples by combining different tissue references with specified proportions.
The resulting mixed samples are simulated by first generating random proportions for each sample using
the Dirichlet distribution, then combining pure tissue reference panels with noise. This function can be
used to simulate gene expression or methylation data where the true proportions of tissue types are
mixed with some noise.
}
\examples{
# Example of generating mixed samples
N_sample = 10
logpure_base = matrix(runif(100), nrow = 10, ncol = 10)  # Random pure tissue base
logpure_sd = matrix(runif(100) * 0.1, nrow = 10, ncol = 10)  # Random standard deviation for pure tissue
result = getSampleMix(N_sample, logpure_base, logpure_sd)
head(result$obs.Y)  # View the generated mixed samples
}
