% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBatch.R
\name{removeBatch}
\alias{removeBatch}
\title{Title removeBatch}
\usage{
removeBatch(dat.raw, Jaff.blood.ref)
}
\arguments{
\item{dat.raw}{A numeric matrix or data frame containing the raw data. The rows represent features (e.g., genes),
and the columns represent samples. The samples in \code{dat.raw} should match the samples in \code{Jaff.blood.ref}.}

\item{Jaff.blood.ref}{A numeric matrix or data frame containing reference data to which batch effects will be removed.
The rows should correspond to the same features as in \code{dat.raw}. The columns 3 to 8 in \code{Jaff.blood.ref}
are used for the batch effect adjustment.}
}
\value{
A list with two components:
\item{Y.raw}{A matrix containing the gene expression data with the batch effects removed. It is the adjusted data from \code{dat.raw}.}
\item{pure}{A matrix containing the reference data with the batch effects removed. It corresponds to the adjusted data from \code{Jaff.blood.ref}.}
}
\description{
#' Remove Batch Effects from Data Using ComBat
}
\details{
This function removes batch effects from gene expression data by applying the ComBat algorithm.
It uses a reference dataset (\code{Jaff.blood.ref}) to remove systematic differences between
the two data sets (\code{dat.raw} and \code{Jaff.blood.ref}).
}
