% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vca.R
\name{vca}
\alias{vca}
\title{Title vca}
\usage{
vca(R, p, SNR = NULL, verbose = F)
}
\arguments{
\item{R}{A numeric matrix of size L x N, where L is the number of genes (rows) and N is the number of samples (columns).
Each element represents the expression level of a gene in a sample.}

\item{p}{An integer specifying the number of components to extract.}

\item{SNR}{Optional. A numeric value for the Signal-to-Noise Ratio. If provided, the analysis will adapt based on the SNR threshold.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is \code{FALSE}.}
}
\value{
A numeric matrix of size L x p, containing the extracted components from the input data matrix R.
The columns of the returned matrix represent the variance components.
}
\description{
Variance Component Analysis (VCA)
}
\details{
This function performs Variance Component Analysis (VCA) to extract relevant components from the input data matrix.
VCA is often used for decomposing variance in gene expression data to identify principal components or sources of variation.
}
\examples{
# Example usage of the vca function
R <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Random example data (100 genes x 10 samples)
p <- 3  # Number of components to extract
vca_results <- vca(R, p)
}
